requirements.txt
fastapi[standard]>=0.111
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
transformers>=4.46
sentence-transformers
torch>=2.0.0
scikit-learn
joblib
pdfplumber>=0.11.0
python-docx>=1.1.0
langdetect>=1.0.9
pdf2image>=1.17.0
pytesseract>=0.3.10
pillow>=10.0.0
datasets>=4.0
httpx>=0.27
qdrant-client>=1.9.0
celery>=5.3.0
redis>=5.0.0
pytest>=8.0.0
hypothesis>=6.100.0
fpdf2>=2.7.0
